refactor: remove DataTypeSelect component and update related views and tests
